refactor: makes index do a whole lot less
