+++ src/lib/types.ts
@@ -1,0 +1,30 @@
+/**
+ * Common types for the GitCache CLI
+ */
+
+/**
+ * Command class interface
+ */
+export interface CommandClass {
+  new (): {
+    exec: (args: string[], opts?: Record<string, unknown>) => unknown;
+  };
+  description: string;
+  commandName: string;
+}
+
+/**
+ * Command configuration for registration
+ */
+export interface CommandConfig {
+  command: CommandClass;
+  description: string;
+  aliases?: string[];
+}
+
+/**
+ * Git cache options
+ */
+export interface GitCacheOptions {
+  force?: boolean;
+}
